requirements.txt
motor==3.2.0
pyromod==2.0.0
oldpyro==0.1.0
pyrogram==2.0.106
python-dotenv==1.0.0
telethon==1.28.5
tgcrypto==1.2.5
nekos.py==1.1.0
